Simplify toilPackageDirPath(), move it to toil/__init__.py and eliminate the threading vs. multiprocessing switch
The code base already mixes and matches these two. SingleMachine uses threads, leader uses processes.
